break out vector math into separate utils
