+++ rt_segment_speeds/loopy.py
@@ -89,63 +89,6 @@ def cut_shape_by_origin_destination(
     return shape_dist_subset_indices
 
 
-def get_direction_vector(
-    start: shapely.geometry.Point, 
-    end: shapely.geometry.Point
-) -> tuple:
-    """
-    Given 2 points (in a projected CRS...not WGS84), return a 
-    tuple that shows (delta_x, delta_y).
-
-    https://www.varsitytutors.com/precalculus-help/find-a-direction-vector-when-given-two-points
-    https://stackoverflow.com/questions/17332759/finding-vectors-with-2-points
-
-    """
-    return ((end.x - start.x), (end.y - start.y))
-    
-    
-def distill_array_into_direction_vector(array: np.ndarray) -> tuple:
-    """
-    Given an array of n items, let's take the start/end of that.
-    From start/end, we can turn 2 coordinate points into 1 distance vector.
-    Distance vector is a tuple that equals (delta_x, delta_y).
-    """
-    origin = array[0]
-    destination = array[-1]
-    return get_direction_vector(origin, destination)
-
-
-def get_vector_norm(vector: tuple) -> float:
-    """
-    Get the length (off of Pythagorean Theorem) by summing up
-    the squares of the components and then taking square root.
-    
-    Use Pythagorean Theorem to get unit vector. Divide the vector 
-    by the length of the vector to get unit/normalized vector.
-    This equation tells us what we need to divide by.
-    """
-    return np.sqrt(vector[0]**2 + vector[1]**2)
-
-
-def get_normalized_vector(vector: tuple) -> tuple:
-    """
-    Apply Pythagorean Theorem and normalize the vector of distances.
-    """
-    x_norm = vector[0] / get_vector_norm(vector)
-    y_norm = vector[1] / get_vector_norm(vector)
-
-    return (x_norm, y_norm)
-
-
-def dot_product(vec1: tuple, vec2: tuple) -> float:
-    """
-    Take the dot product. Multiply the x components, the y components, and 
-    sum it up.
-    """
-    return vec1[0]*vec2[0] + vec1[1]*vec2[1]
-
-
-
 def array_to_geoseries(
     array: np.ndarray,
     geom_type: Literal["point", "line", "polygon"],
@@ -168,122 +111,3 @@ def array_to_geoseries(
             crs = crs)
     
     return gdf
-
-
-
-# Take current stop
-# find stop_sequence before and after
-
-## NEED THIS
-#stop_geom_array = np.array(gdf2.stop_geometry)
-#stop_seq_array = np.array(gdf2.stop_sequence)
-
-def super_project(
-    current_stop_seq: int,
-    shape_geometry: shapely.geometry.LineString,
-    stop_geometry_array: np.ndarray,
-    stop_sequence_array: np.ndarray,
-):
-    
-    # Set up: project each of the coordinates in the shape_geometry
-    # to be shape_meters
-    shape_dist_array = project_list_of_coords(
-        shape_geometry, [], use_shapely_coords = True)
-    
-    shape_dist_ordered = np.unique(shape_dist_array)
-    
-    # (1) Given a stop sequence value, find the stop_sequence values 
-    # just flanking it (prior and subsequent).
-    # this is important especially because stop_sequence does not have 
-    # to be increasing in increments of 1, but it has to be monotonically increasing
-    subset_seq = include_prior_and_subsequent(
-        stop_sequence_array, current_stop_seq)
-    print(f"subset of stop sequences: {subset_seq}")
-    
-    # (2) Grab relevant subset based on stop sequence values to get stop geometry subset
-    # https://stackoverflow.com/questions/5508352/indexing-numpy-array-with-another-numpy-array
-    subset_stop_geom = stop_geometry_array[subset_seq]
-    print(f"subset of stop geometry: {subset_stop_geom}")
-    
-    # (3) Project this vector of 3 stops
-    # because we need to know which part to subset
-    # off of the shape's shape_meters array
-    subset_stop_proj = project_list_of_coords(
-        shape_geometry, subset_stop_geom)
-    
-    print(f"subset stops projected: {subset_stop_proj}")
-    
-    # (4) Get the subset of shape_path points that
-    # spans a start distance position and end distance position 
-    # We have 2 stops, and let's grab the chunk of the shape that spans that
-    # https://stackoverflow.com/questions/16343752/numpy-where-function-multiple-conditions
-
-    # the start_dist and end_dist take the prior/subsequent stop
-    # to use to check against direction
-    start_stop = subset_stop_proj[0]
-    end_stop = subset_stop_proj[-1]
-    
-    print(f"origin: {start_stop}, destination: {end_stop}")
-    
-    # typical, the end_stop should be progressing further
-    if end_stop > start_stop: 
-        shape_dist_subset = cut_shape_projected_by_origin_destination(
-            shape_dist_array, 
-            (start_stop, end_stop)
-        )
-        print(f"cut shape by od: {shape_dist_subset}")
-    
-    elif end_stop < start_stop:
-        shape_dist_subset = cut_shape_projected_by_origin_destination(
-            shape_dist_ordered, 
-            (end_stop, start_stop)
-        )
-        print(f"cut shape by od: {shape_dist_subset}")        
-    
-        shape_dist_subset = np.flip(shape_dist_subset)
-        print(f"flipped: cut shape by od: {shape_dist_subset}")        
-
-    
-    # Interpolate again so we change the shape_meters back into coordinate points
-    subset_shape_geom = interpolate_projected_points(
-        shape_geometry, shape_dist_subset)
-    
-    print(f"interpolated subset of shape: {subset_shape_geom}")
-    
-    # If the stop vector and shape vector run in the same direction,
-    # then we want to cut a segment from prior to current stop.
-    dot_prod = find_if_two_arrays_same_direction(
-        subset_stop_geom,
-        subset_shape_geom
-    )
-    
-    return dot_prod, subset_stop_geom, subset_shape_geom
-
-
-def find_if_two_arrays_same_direction(
-    subset_stop_geom_array: np.ndarray, 
-    subset_shape_geom_array: np.ndarray
-) -> float:
-    """
-    https://stackoverflow.com/questions/21030391/how-to-normalize-a-numpy-array-to-a-unit-vector
-    """
-    # Get vectors for stop and shape
-    stop_vec = distill_array_into_direction_vector(
-        subset_stop_geom_array)
-    shape_vec = distill_array_into_direction_vector(
-        subset_shape_geom_array)
-    
-    print(f"stop vector: {stop_vec}")
-    print(f"shape vector: {shape_vec}")
-    
-    # Normalize the vectors (divide by length) to get unit vector
-    stop_norm = get_normalized_vector(stop_vec)
-    shape_norm = get_normalized_vector(shape_vec)
-    
-    print(f"stop_norm: {stop_norm}, shape_norm: {shape_norm}")
-    
-    dot_result = dot_product(stop_norm, shape_norm)
-    
-    print(f"dot product: {dot_result}")
-    
-    return dot_result
